requirements/requirements.txt
boto3==1.34.79
cachetools==5.3.3
celery==5.3.6
celery-types==0.22.0
colorlog==6.8.2
dj_database_url==2.1.0
Django==5.0.4
django-cors-headers==4.3.1
django-split-settings==1.3.1 # https://pypi.org/project/django-split-settings/
django-storages==1.14.2
django_redis==5.4.0
djangorestframework==3.15.1
djangorestframework_simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==21.2.0
psycopg[c]==3.1.18
pyjwt==2.8.0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
sentry-sdk==1.45.0
python-decouple==3.8
